+++ src/CoreEx.Newtonsoft/Json/EventDataSerializer.cs
@@ -16,9 +16,9 @@ namespace CoreEx.Newtonsoft.Json
         /// </summary>
         /// <param name="jsonSerializer">The <see cref="IJsonSerializer"/>.</param>
         /// <param name="eventDataFormatter">The <see cref="Events.EventDataFormatter"/>.</param>
-        public EventDataSerializer(IJsonSerializer jsonSerializer, EventDataFormatter? eventDataFormatter = null) : base(jsonSerializer, eventDataFormatter)
+        public EventDataSerializer(IJsonSerializer? jsonSerializer = null, EventDataFormatter? eventDataFormatter = null) : base(jsonSerializer ?? new JsonSerializer(), eventDataFormatter)
         {
-            if (jsonSerializer is not CoreEx.Newtonsoft.Json.JsonSerializer)
+            if (JsonSerializer is not CoreEx.Newtonsoft.Json.JsonSerializer)
                 throw new ArgumentException($"The {nameof(IJsonSerializer)} instance must be of Type '{typeof(CoreEx.Newtonsoft.Json.JsonSerializer).FullName}'.", nameof(jsonSerializer));
         }
     }